refactor: remove old iskp
