+++ packages/cryptoplease/lib/core/payments/split_key_payments/transaction/creators/cp_tx_creator.dart
@@ -2,7 +2,7 @@ import 'package:cryptoplease/config.dart';
 import 'package:cryptoplease/core/accounts/bl/account.dart';
 import 'package:cryptoplease/core/payments/split_key_payments/transaction/tx_creator.dart';
 import 'package:cryptoplease/core/resign_tx.dart';
-import 'package:cryptoplease/features/incoming_split_key_payment/bl/tx_processor.dart';
+import 'package:cryptoplease/features/incoming_split_key_payments/presentation/first_part_ready_screen.dart';
 import 'package:cryptoplease/features/outgoing_transfer/bl/outgoing_payment.dart';
 import 'package:cryptoplease_api/cryptoplease_api.dart';
 import 'package:dfunc/dfunc.dart';
