+++ packages/cryptoplease/lib/features/activities/bl/activity.dart
@@ -1,10 +1,7 @@
 import 'package:collection/collection.dart';
-import 'package:cryptoplease/features/incoming_split_key_payment/bl/models.dart';
-import 'package:cryptoplease/features/incoming_split_key_payment/bl/repository.dart';
 import 'package:cryptoplease/features/outgoing_transfer/bl/outgoing_payment.dart';
 import 'package:cryptoplease/features/outgoing_transfer/bl/repository.dart';
 import 'package:cryptoplease/features/payment_request/bl/repository.dart';
-import 'package:dfunc/dfunc.dart';
 import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:rxdart/rxdart.dart';
@@ -13,10 +10,6 @@ part 'activity.freezed.dart';
 
 @freezed
 class Activity with _$Activity {
-  factory Activity.splitKeyIncoming({
-    required SplitKeyIncomingFirstPart firstPart,
-  }) = SplitKeyIncomingActivity;
-
   factory Activity.outgoingTransfer(OutgoingTransfer transfer) =
       OutgoingTransferActivity;
 
@@ -27,8 +20,7 @@ class Activity with _$Activity {
 
   const Activity._();
 
-  DateTime? get created => this.map(
-        splitKeyIncoming: always(null),
+  DateTime? get created => map(
         outgoingTransfer: (a) => a.transfer.created,
         paymentRequest: (a) => a.created,
       );
@@ -36,30 +28,21 @@ class Activity with _$Activity {
 
 Stream<IList<Activity>> watchActivities({
   required OutgoingTransferRepository outgoingRepository,
-  required SplitKeyIncomingRepository incomingRepository,
   required PaymentRequestRepository paymentRequestRepository,
 }) {
   final outgoing = outgoingRepository
       .watchPayments()
       .map((ps) => ps.map(Activity.outgoingTransfer));
 
-  final incoming = incomingRepository.watch().map(
-        (v) => v == null
-            ? <Activity>[]
-            : [Activity.splitKeyIncoming(firstPart: v)],
-      );
-
   final paymentRequests = paymentRequestRepository.watchAllIds().map(
         (ids) =>
             ids.map((p) => Activity.paymentRequest(p.item1, created: p.item2)),
       );
 
-  return CombineLatestStream.combine3(
-    incoming,
+  return CombineLatestStream.combine2(
     paymentRequests,
     outgoing,
-    (Iterable<Activity> i, Iterable<Activity> p, Iterable<Activity> o) => [
-      ...i,
+    (Iterable<Activity> p, Iterable<Activity> o) => [
       // ignore: avoid-non-null-assertion, outgoing payments and requests should have created field
       ...IList([...p, ...o]).sortedBy((e) => e.created!).reversed,
     ].toIList(),
