--- packages/cryptoplease/lib/features/incoming_split_key_payment/bl/event.dart
@@ -1,22 +1,0 @@
-part of 'bloc.dart';
-
-@freezed
-class SplitKeyIncomingPaymentEvent with _$SplitKeyIncomingPaymentEvent {
-  const factory SplitKeyIncomingPaymentEvent.firstPartAdded({
-    required SplitKeyIncomingFirstPart firstPart,
-  }) = FirstPartAdded;
-
-  const factory SplitKeyIncomingPaymentEvent.secondPartAdded({
-    required SplitKeySecondLink value,
-  }) = SecondPartAdded;
-
-  const factory SplitKeyIncomingPaymentEvent.paymentRequested({
-    required String recipient,
-  }) = PaymentRequested;
-
-  const factory SplitKeyIncomingPaymentEvent.retried({
-    required String recipient,
-  }) = Retried;
-
-  const factory SplitKeyIncomingPaymentEvent.cleared() = PaymentCleared;
-}
