--- packages/cryptoplease/lib/features/incoming_split_key_payment/bl/repository.dart
@@ -1,7 +1,0 @@
-import 'package:cryptoplease/features/incoming_split_key_payment/bl/models.dart';
-
-abstract class SplitKeyIncomingRepository {
-  Stream<SplitKeyIncomingFirstPart?> watch();
-  Future<void> save(SplitKeyIncomingFirstPart firstPart);
-  Future<void> clear();
-}
