+++ packages/cryptoplease/test/golden/golden_test.dart
@@ -1,7 +1,5 @@
-import 'package:cryptoplease/features/incoming_split_key_payment/presentation/components/success_content.dart';
 import 'package:cryptoplease/features/onboarding/presentation/sign_up/get_started_screen.dart';
 import 'package:cryptoplease/features/onboarding/presentation/sign_up/no_email_and_password_screen.dart';
-import 'package:flutter_test/flutter_test.dart';
 
 import 'utils.dart';
 
@@ -12,11 +10,4 @@ void main() {
     'No email and password screen',
     const NoEmailAndPasswordScreen(),
   );
-
-  group('SplitKeyIncomingPaymentScreen', () {
-    testGoldensWidget(
-      'SplitKeyIncomingPaymentScreen success',
-      const SuccessContent(),
-    );
-  });
 }
